Implementación de la paginación.
Se implemento la paginación con el
-pageNumber
-pageSize
-sortBy
-sortDir

+++ src/main/java/com/ronal/blog/service/impl/PublicationServiceImpl.java
@@ -1,9 +1,10 @@
 package com.ronal.blog.service.impl;
 
-import com.ronal.blog.dto.PublicationDTO;
-import com.ronal.blog.dto.PublicationResponse;
 import com.ronal.blog.dao.PublicationDAO;
+import com.ronal.blog.dto.PublicationDTO;
+import com.ronal.blog.dto.PublicationResponseDTO;
 import com.ronal.blog.dto.ResponseDTO;
+import com.ronal.blog.exceptions.RequestException;
 import com.ronal.blog.mapper.PublicationMapper;
 import com.ronal.blog.repository.PublicationRepository;
 import com.ronal.blog.service.PublicationService;
@@ -12,11 +13,13 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
+
+import static com.ronal.blog.util.Constantes.BAD_REQUEST_TITLE;
 
 @RequiredArgsConstructor
 @Service
@@ -27,56 +30,21 @@ public class PublicationServiceImpl implements PublicationService {
 
 
     @Override
+    public PublicationResponseDTO listPublications(int pageNumber, int pageSize, String sortBy, String sortDir) {
+        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
+        Pageable pageable = PageRequest.of(pageNumber, pageSize, sort);
+        Page<PublicationDAO> publicationDAOS = publicationRepository.findAll(pageable);
+        List<PublicationDAO> listaDePublicaciones = publicationDAOS.getContent();
+        List<PublicationDTO> listaSalida = publicationMapper.listDaoToListDto(listaDePublicaciones);
+        return PublicationResponseDTO.builder().contenido(listaSalida).pageNumber(publicationDAOS.getNumber()).pageSize(publicationDAOS.getSize()).totalElementos(publicationDAOS.getTotalElements()).totalPaginas(publicationDAOS.getTotalPages()).build();
+    }
+
+    @Override
     public ResponseDTO<PublicationDTO> save(PublicationDTO dto) {
-
         PublicationDAO publicationDAO = publicationMapper.publicationDtoToPublicationDao(dto);
-
         publicationRepository.save(publicationDAO);
-
-        PublicationDTO salida  = publicationMapper.publicationDaoToPublicationDto(publicationDAO);
-
-        return ResponseDTO.<PublicationDTO>builder()
-                .success(Boolean.TRUE)
-                .mensaje("REGISTRADO CORRECTAMENTE")
-                .response(salida).build();
-    }
-
-    @Override
-    public List<PublicationDTO> listaRonal() {
-        List<PublicationDAO> listaDAO = publicationRepository.findAll();
-        return  publicationMapper.lstToDto(listaDAO);
-    }
-
-    @Override
-    public List<PublicationDTO> list(int numeroDePagina, int medidaDePagina) {
-        Pageable pageable = PageRequest.of(numeroDePagina,medidaDePagina);
-
-        Page<PublicationDAO> publicacions = publicationRepository.findAll(pageable);
-
-        List<PublicationDAO> listaDePublicaciones = publicacions.getContent();
-       return listaDePublicaciones.stream().map(publicationDAO -> mapearDTO(publicationDAO)).collect(Collectors.toList());
-    }
-
-    @Override
-    public PublicationResponse listar(int numeroDePagina, int medidaDePagina, String ordenarPor, String sortDir) {
-        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(ordenarPor).ascending():Sort.by(ordenarPor).descending();
-        Pageable pageable = PageRequest.of(numeroDePagina,medidaDePagina,sort);
-
-        Page<PublicationDAO> publicacions = publicationRepository.findAll(pageable);
-
-        List<PublicationDAO> listaDePublicaciones = publicacions.getContent();
-
-        List<PublicationDTO> contenido =  listaDePublicaciones.stream().map(publicationDAO -> mapearDTO(publicationDAO)).collect(Collectors.toList());
-
-        PublicationResponse publicationResponse = new PublicationResponse();
-        publicationResponse.setContenido(contenido);
-        publicationResponse.setNumeroDePagina(publicacions.getNumber());
-        publicationResponse.setMedidaDePagina(publicacions.getSize());
-        publicationResponse.setTotalElementos(publicacions.getTotalElements());
-        publicationResponse.setTotalPaginas(publicacions.getTotalPages());
-        publicationResponse.setUltima(publicacions.isLast());
-
-        return publicationResponse;
+        PublicationDTO salida = publicationMapper.publicationDaoToPublicationDto(publicationDAO);
+        return ResponseDTO.<PublicationDTO>builder().success(Boolean.TRUE).mensaje("REGISTRADO CORRECTAMENTE").response(salida).build();
     }
 
     @Override
@@ -89,7 +57,7 @@ public class PublicationServiceImpl implements PublicationService {
     public PublicationDTO update(PublicationDTO dto, Long id) {
 
         PublicationDAO publicationDAO = publicationRepository.findById(id).orElse(null);
-        if (Objects.isNull(publicationDAO)){
+        if (Objects.isNull(publicationDAO)) {
             return null;
         }
 
@@ -105,55 +73,9 @@ public class PublicationServiceImpl implements PublicationService {
     @Override
     public void delete(Long id) {
         PublicationDAO publicationDAO = publicationRepository.findById(id).orElse(null);
-
+        if (Objects.isNull(publicationDAO)) {
+            throw new RequestException(BAD_REQUEST_TITLE, "C-50", HttpStatus.BAD_REQUEST);
+        }
         publicationRepository.delete(publicationDAO);
     }
-
-
-    //ENTITY => DTO
-    @Deprecated
-    public PublicationDTO mapearDTO(PublicationDAO publicationDAO){
-        PublicationDTO publicationDTO = new PublicationDTO();
-        publicationDTO.setIdPublication(publicationDAO.getIdPublication());
-        publicationDTO.setTitle(publicationDAO.getTitle());
-        publicationDTO.setDescription(publicationDAO.getDescription());
-        publicationDTO.setContent(publicationDAO.getContent());
-        return publicationDTO;
-    }
-    @Deprecated
-    public PublicationDAO mapearEntity(PublicationDTO dto) {
-        PublicationDAO publicationDAO = new PublicationDAO();
-        publicationDAO.setTitle(dto.getTitle());
-        publicationDAO.setDescription(dto.getDescription());
-        publicationDAO.setContent(dto.getContent());
-        return publicationDAO;
-    }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
